updated Objects: Player new const. Team: find()
Added a new constructor for the Player objects.
You can now construct a player with just a name (string).
The position is currently not being used (may not get used at all, and
will be removed later if found not useful).

The Team object has two new methods for linearly searching for
a player either by a string name or by a player object (Must be a
literal object: IE, objects must be the same literal object).