Added Player-to-Team dictionary
Added the dictionary within TeamCalculator.java which can be used to map
a player to a team. This dictionary is initialized and filled within the
randomize() method. Note that the dictionary will be a new dictionary
every time you call the randomize() method (since each player  has the
potential to change teams). The Dictionary is implemented as a HashTable
who's layout is:
Player<Object> (Key) : Team<Object> (Value)

The Dictionary is tested within the TeamCalculatorTest.java file. All
J-Unit test cases pass.